ai-service/requirements.txt
fastapi
uvicorn
python-multipart=
pdfplumber
PyMuPDF
pypdf2
langchain
langchain_community
sentence-transformers
chromadb
python-dotenv
langchain-google-genai
